requirements.txt
fastapi
uvicorn
bcrypt
python-jose[cryptography]
passlib[bcrypt]
pydantic
python-multipart
